pyproject.toml
# --
# Copyright (c) 2008-2025 Net-ng.
# All rights reserved.
#
# This software is licensed under the BSD License, as described in
# the file LICENSE.txt, which you should have received as part of
# this distribution.
# --

[build-system]
requires = ['setuptools', 'wheel', 'setuptools_scm']
build-backend = 'setuptools.build_meta'

[tool.setuptools_scm]

[project]
name = 'nagare-services-router-rules'
dynamic = ['version']
description = 'REST url router service'
authors = [
    {name = 'Net-ng'},
    {email = 'contact@net-ng.com'}
]
license = {file = 'LICENSE.txt'}
requires-python = '>=3.7'
dependencies = [
    'webob',
    "legacy-cgi ; python_full_version>='3.13.0a1'",
    'nagare-peak-rules',
    'nagare-services',
    'nagare-server-http'
]
[project.readme]
file = 'README.rst'
content-type = 'text/x-rst'

[project.optional-dependencies]
dev = [
    'sphinx',
    'sphinx_rtd_theme',
    'pre-commit',
    'ruff',
    'pytest',
    'twine'
]

[project.urls]
Homepage = 'https://nagare.org'
Documentation = 'https://nagare.org/doc'
Source = 'https://github.com/nagareproject/services-router-rules'
Tracker = 'https://github.com/nagareproject/services-router-rules/issues'

[tool.ruff]
src = ['src']
line-length = 120
format.quote-style = 'single'

[tool.ruff.lint]
select = ['C4', 'COM', 'D', 'E', 'ERA', 'F', 'I', 'ISC', 'INP', 'PIE', 'Q', 'S', 'SIM', 'TID', 'W', 'YTT']
ignore = ['COM812', 'D10', 'ISC001']
isort.length-sort = true
pydocstyle.convention = 'google'
flake8-quotes.inline-quotes = 'single'

[tool.pytest.ini_options]
testpaths = ['tests']
